refactor: move keyboard shortcuts to useKeyboardNavigation hook
